backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
async-generator==1.10
attrs==25.4.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.38.0
azure-storage-blob==12.28.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
Django==5.2.10
django-cors-headers==4.9.0
django-ninja==1.5.3
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
jiter==0.12.0
msrest==0.7.1
mypy_extensions==1.1.0
oauthlib==3.3.1
openai==2.15.0
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
outcome==1.3.0.post0
packaging==26.0
playwright==1.57.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
PyJWT==2.11.0
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
selenium==4.40.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.5
tqdm==4.67.1
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
zipp==3.23.0
